+++ src/core/login.ts
@@ -31,7 +31,7 @@ export async function login(
   const logConfig = options.logConfig as Partial<LogConfig> | undefined;
   const logger = new Logger(logConfig);
   
-  logger.printBanner('0.6.3');
+  logger.printBanner('0.6.6');
   logger.info(logger.getMessage('loginStart'));
 
   try {